Address flex attention review feedback
